优化代码

+++ points/src/main/java/com/dili/points/domain/PointsRule.java
@@ -1,17 +1,13 @@
 package com.dili.points.domain;
 
 import com.dili.ss.domain.annotation.Like;
-import com.dili.ss.domain.annotation.Operator;
 import com.dili.ss.dto.IBaseDomain;
 import com.dili.ss.metadata.FieldEditor;
 import com.dili.ss.metadata.annotation.EditMode;
 import com.dili.ss.metadata.annotation.FieldDef;
-import java.util.Date;
+
 import javax.persistence.*;
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import java.util.Date;
 
 /**
  * 由MyBatis Generator工具自动生成
@@ -68,9 +64,9 @@ public interface PointsRule extends IBaseDomain {
     @Column(name = "`computing_parameter`")
     @FieldDef(label="计算参数")
     @EditMode(editor = FieldEditor.Number, required = false)
-    Float getComputingParameter();
+    Double getComputingParameter();
 
-    void setComputingParameter(Float computingParameter);
+    void setComputingParameter(Double computingParameter);
 
     @Column(name = "`created_id`")
     @FieldDef(label="创建人")
